+++ src/main/resources/templates/search/search-form.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html lang="en"
+      xmlns:th="http://www.thymeleaf.org">
+<head th:replace="theme-fragments.html :: head"></head>
+<body>
+
+
+
+
+<div id="wrapper">
+
+    <!-- Header -->
+    <header th:replace="theme-fragments.html :: header"></header>
+
+    <!-- sidebar -->
+    <div th:replace="theme-fragments.html :: side_bar"></div>
+
+    <!-- Main Contents -->
+    <div class="main_content">
+        <div class="mcontainer">
+            <form th:object="${searchForm}" method="post" th:action="@{/search}">
+                <div class="grid grid-cols-3 gap-4">
+                    <div>
+                        <select class="bg-gray-200" th:field="*{searchType}">
+                            <option value="스터디">스터디</option>
+                            <option value="커뮤니티">커뮤니티</option>
+                            <option value="장터">장터</option>
+                            <option value="학생회">학생회 공지 및 행사</option>
+                        </select>
+                    </div>
+                    <div class="col-span-2">
+                        <input class="uk-input bg-gray-200 pl-5" placeholder="검색어를 입력해주세요." th:field="*{keyword}">
+                    </div>
+                </div>
+            </form>
+            <!--<th:block th:if="${searchType}">
+                <div class="flex justify-between relative md:mb-4 mb-3">
+                    <a class="text-2xl font-semibold"><a th:text="${searchType}"></a>내 검색 결과</a>
+                </div>
+            </th:block>
+            <th:block th:if="${search_result}">
+                <div th:each="result : ${search_result}">
+                    <div class="divide-y -mt-3 card px-5 py-2 mb-7">
+                        <div class="flex sm:flex-row flex-col sm:space-x-4 py-4 relative w-full">
+                            <div class="flex-1 relative md:mt-0">
+                                <a class="text-xl font-semibold leading-6" th:text="${result.title}"></a>
+                                <div class="flex space-x-2 items-center text-sm mt-2">
+                                    <div>공지등록일</div>
+                                    <div>·</div>
+                                    <div><a></a>명 조회</div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </th:block>-->
+        </div>
+    </div>
+
+</div>
+
+<script th:replace="theme-fragments.html :: form-validation"></script>
+<script type="application/javascript">
+    $( document ).ready( function() {
+
+    });
+</script>
+
+</body>
+</html>
