+++ src/main/resources/templates/theme-fragments.html
@@ -319,9 +319,10 @@
             </div>
 
             <!-- search icon for mobile -->
-            <div class="header-search-icon" uk-toggle="target: #wrapper ; cls: show-searchbox"> </div>
-            <div class="header_search"><i class="uil-search-alt"></i>
-                <input value="" type="text" class="form-control" placeholder="검색어를 입력해주세요." autocomplete="off">
+            <!--<div class="header-search-icon" uk-toggle="target: #wrapper ; cls: show-searchbox"> </div>
+            <div class="header_search">
+                <i class="uil-search-alt"></i>
+                <input name="keyword" type="text" class="form-control" placeholder="검색어를 입력해주세요." autocomplete="off">
                 <div uk-drop="mode: click" class="header_search_dropdown">
                     <h4 class="search_title"> 최근 검색어 </h4>
                     <ul>
@@ -332,11 +333,19 @@
                         </li>
                     </ul>
                 </div>
-            </div>
+            </div>-->
 
             <div class="right_side">
 
                 <div class="header_widgets">
+
+                    <!-- Search -->
+                    <a th:href="@{/search}" class="is_icon" uk-tooltip="title: 검색">
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                            <path d="M9 9a2 2 0 114 0 2 2 0 01-4 0z"></path>
+                            <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-13a4 4 0 00-3.446 6.032l-2.261 2.26a1 1 0 101.414 1.415l2.261-2.261A4 4 0 1011 5z" clip-rule="evenodd"></path>
+                        </svg>
+                    </a>
 
                     <a href="#" class="is_icon" uk-tooltip="title: 알림">
                         <svg fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path d="M10 2a6 6 0 00-6 6v3.586l-.707.707A1 1 0 004 14h12a1 1 0 00.707-1.707L16 11.586V8a6 6 0 00-6-6zM10 18a3 3 0 01-3-3h6a3 3 0 01-3 3z"></path></svg>
